requirements.txt
langchain
langchain-openai
openai
slack-bolt
pyyaml
